- Fixed issues for adding new data for patient does not have same UID & PID - Fixed possibility issues divide by zero while adding weight & height, which come from generate bmi value

+++ apps/data_app/operations.py
@@ -18,4 +18,4 @@ def get_observation_most_common_value(component, effective_datetime):
         if most_recent_value.exists():
             return float(most_recent_value.first().value_quantity)
         else:
-            return float(0)
+            return float(1)
